pyproject.toml
[project]
name = "anything-to-podcast"
version = "0.1.0"
description = "Convert various sources to Chinese podcast episodes"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "pymupdf>=1.24",
    "arxiv>=2.1",
    "edge-tts>=6.1",
    "openai>=1.0",
    "pyyaml>=6.0",
    "feedgen>=1.0",
]

[project.scripts]
a2p = "main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
